# perf: improve setTimestamp, setTime, setDate performance