Avoid Calendar cloning and Integer.toString in setCalendar. This improves setTimestamp from 0.8us to 0.2us

Before
```
Benchmark                                                      Mode  Cnt    Score     Error   Units
BindTimestamp.timestampCal                                     avgt    5  754,114 ±  91,625   ns/op
BindTimestamp.timestampCal:·gc.alloc.rate                      avgt    5  931,234 ± 111,715  MB/sec
BindTimestamp.timestampCal:·gc.alloc.rate.norm                 avgt    5  736,007 ±   0,061    B/op
BindTimestamp.timestampCal:·gc.churn.PS_Eden_Space             avgt    5  934,073 ±  86,972  MB/sec
BindTimestamp.timestampCal:·gc.churn.PS_Eden_Space.norm        avgt    5  738,410 ±  34,828    B/op
BindTimestamp.timestampCal:·gc.churn.PS_Survivor_Space         avgt    5    0,405 ±   0,189  MB/sec
BindTimestamp.timestampCal:·gc.churn.PS_Survivor_Space.norm    avgt    5    0,320 ±   0,170    B/op
BindTimestamp.timestampCal:·gc.count                           avgt    5  113,000            counts
BindTimestamp.timestampCal:·gc.time                            avgt    5   61,000                ms
BindTimestamp.timestampLocal                                   avgt    5  661,347 ±  30,940   ns/op
BindTimestamp.timestampLocal:·gc.alloc.rate                    avgt    5  346,037 ±  16,301  MB/sec
BindTimestamp.timestampLocal:·gc.alloc.rate.norm               avgt    5  240,007 ±   0,056    B/op
BindTimestamp.timestampLocal:·gc.churn.PS_Eden_Space           avgt    5  346,885 ±  39,791  MB/sec
BindTimestamp.timestampLocal:·gc.churn.PS_Eden_Space.norm      avgt    5  240,608 ±  27,597    B/op
BindTimestamp.timestampLocal:·gc.churn.PS_Survivor_Space       avgt    5    0,262 ±   0,183  MB/sec
BindTimestamp.timestampLocal:·gc.churn.PS_Survivor_Space.norm  avgt    5    0,182 ±   0,136    B/op
BindTimestamp.timestampLocal:·gc.count                         avgt    5   74,000            counts
BindTimestamp.timestampLocal:·gc.time                          avgt    5   40,000                ms
```

After:

```
Benchmark                                                      Mode  Cnt    Score    Error   Units
BindTimestamp.timestampCal                                     avgt    5  214,056 ± 15,606   ns/op
BindTimestamp.timestampCal:·gc.alloc.rate                      avgt    5  463,340 ± 33,136  MB/sec
BindTimestamp.timestampCal:·gc.alloc.rate.norm                 avgt    5  104,002 ±  0,018    B/op
BindTimestamp.timestampCal:·gc.churn.PS_Eden_Space             avgt    5  463,675 ± 59,896  MB/sec
BindTimestamp.timestampCal:·gc.churn.PS_Eden_Space.norm        avgt    5  104,087 ± 12,810    B/op
BindTimestamp.timestampCal:·gc.churn.PS_Survivor_Space         avgt    5    0,287 ±  0,428  MB/sec
BindTimestamp.timestampCal:·gc.churn.PS_Survivor_Space.norm    avgt    5    0,064 ±  0,093    B/op
BindTimestamp.timestampCal:·gc.count                           avgt    5   87,000           counts
BindTimestamp.timestampCal:·gc.time                            avgt    5   49,000               ms
BindTimestamp.timestampLocal                                   avgt    5  157,395 ±  8,393   ns/op
BindTimestamp.timestampLocal:·gc.alloc.rate                    avgt    5  630,067 ± 33,089  MB/sec
BindTimestamp.timestampLocal:·gc.alloc.rate.norm               avgt    5  104,002 ±  0,013    B/op
BindTimestamp.timestampLocal:·gc.churn.PS_Eden_Space           avgt    5  633,737 ± 43,571  MB/sec
BindTimestamp.timestampLocal:·gc.churn.PS_Eden_Space.norm      avgt    5  104,635 ± 11,285    B/op
BindTimestamp.timestampLocal:·gc.churn.PS_Survivor_Space       avgt    5    0,336 ±  0,427  MB/sec
BindTimestamp.timestampLocal:·gc.churn.PS_Survivor_Space.norm  avgt    5    0,055 ±  0,070    B/op
BindTimestamp.timestampLocal:·gc.count                         avgt    5   89,000           counts
BindTimestamp.timestampLocal:·gc.time                          avgt    5   49,000               ms
```