@davecramer 2015-10-05
I have no idea why the clone was added in the first place. Is there some best practices somewhere which recommends this ?