@vlsi 2015-10-05
I think `clone()` was added to prevent modification of user-provided `Calendar`.
I did not dig much into the spec to check if JDBC driver is allowed to modify `Calendar` that is passed via `setTimestamp(int, Timestamp, Calendar)` kind of calls.

I just made sure user-provided `Calendar` is used in a read-only way: https://github.com/pgjdbc/pgjdbc/pull/379/files#diff-d3ce26f4020b4504cf0a8ef2e019b316R489